Initial commit of core files
These files have been sitting on my personal disk for a while. Algo is
the algorithm itself. Feed is what gets the data from IB. Order is what
reads what algo decided and executes the orders. Master_algo/feed are
multiprocessing scripts that will open multiple threads of algo and
feed, each with their own unique client ID#. Master_feed looks at what
is contained in ./stocks and creates an instance of feed for that
specific ticker. Master_algo will also look at the ./stocks file and
will try looking for the out files of the feed instances. It will create
a unique instance of the algorithms. For ./order.py this master script
isnt nessecary in its current state because it will find the files based
on a search. I do this to limit the number of client connections to IB,
as there is (probably?) a limit. I noticed I could only get ~45 or 50
clients connecting to the API before it started throwing errors.