Major Update
If you're looking for code, look in the subdir v0 for something that
works. I'm reworking alot of this so it feels less... clunky. Even
though I'm probably going to be the only person to ever actually use
this, I want it to feel clean, logical, and intuitive as to how it
works.